+++ Examples/LocationManager/LocalSearchClient/Package.swift
@@ -3,7 +3,7 @@
 import PackageDescription
 
 let package = Package(
-  name: "LocalSearchClient",
+  name: "Common",
   platforms: [
     .iOS(.v13),
     .macOS(.v10_15),
@@ -11,21 +11,25 @@ let package = Package(
     .watchOS(.v6),
   ],
   products: [
-    // Products define the executables and libraries produced by a package, and make them visible to other packages.
     .library(
       name: "LocalSearchClient",
       targets: ["LocalSearchClient"]),
+    .library(
+      name: "Core",
+      targets: ["Core"]),
   ],
   dependencies: [
-    // Dependencies declare other packages that this package depends on.
-    // .package(url: /* package url */, from: "1.0.0"),
     .package(path: "../../../")
   ],
   targets: [
-    // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-    // Targets can depend on other targets in this package, and on products in packages which this package depends on.
     .target(
       name: "LocalSearchClient",
       dependencies: ["ComposableArchitecture"]),
+    .target(
+      name: "Core",
+      dependencies: ["ComposableArchitecture", "ComposableCoreLocation", "LocalSearchClient"]),
+    .testTarget(
+      name: "CoreTests",
+      dependencies: ["Core"]),
   ]
 )
