+++ Examples/LocationManager/LocalSearchClient/Tests/CoreTests/CoreTests.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by Brandon Williams on 5/19/20.
+//
+
+import Foundation
