--- Examples/LocationManager/LocalSearchClient/Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import LocalSearchClientTests
-
-var tests = [XCTestCaseEntry]()
-tests += LocalSearchClientTests.allTests()
-XCTMain(tests)
